+++ crest/Assets/Crest/Crest/Scripts/Shapes/ShapeGerstner.cs
@@ -54,6 +54,14 @@ namespace Crest
         [Tooltip("Multiplier for these waves to scale up/down."), Range(0f, 1f)]
         public float _weight = 1f;
 
+        [Predicated(typeof(ShapeFFT), "IsLocalWaves"), DecoratedField]
+        [Tooltip("If enabled, waves are additive. Otherwise waves are blended.")]
+        public bool _additive = true;
+
+        [Predicated(typeof(ShapeFFT), "IsLocalWaves"), DecoratedField]
+        [Tooltip("Order this input will render.")]
+        public int _queue = 0;
+
         [Tooltip("How much these waves respect the shallow water attenuation setting in the Animated Waves Settings. Set to 0 to ignore shallow water."), SerializeField, Range(0f, 1f)]
         public float _respectShallowWaterAttenuation = 1f;
 
@@ -322,6 +330,12 @@ namespace Crest
             // Seems like shader errors cause this to unbind if I don't set it every frame. Could be an editor only issue.
             _matGenerateWaves.SetTexture(sp_WaveBuffer, _waveBuffers);
 
+            if (_meshForDrawingWaves != null && _matGenerateWavesGeometry != null)
+            {
+                _matGenerateWavesGeometry.SetInt(Helpers.ShaderIDs.s_BlendSrcMode, (int)(_additive ? BlendMode.One : BlendMode.SrcAlpha));
+                _matGenerateWavesGeometry.SetInt(Helpers.ShaderIDs.s_BlendDstMode, (int)(_additive ? BlendMode.One : BlendMode.OneMinusSrcAlpha));
+            }
+
             ReportMaxDisplacement();
 
             // If some cascades have waves in them, generate
@@ -352,6 +366,12 @@ namespace Crest
             {
                 _meshForDrawingWaves = null;
             }
+        }
+
+        // Called by Predicated attribute. Signature must not be changed.
+        bool IsLocalWaves(Component component)
+        {
+            return TryGetComponent<MeshRenderer>(out _) || TryGetComponent<Spline.Spline>(out _);
         }
 #endif
 
@@ -656,6 +676,9 @@ namespace Crest
                 }
 
                 _matGenerateWaves = _matGenerateWavesGeometry;
+
+                _matGenerateWavesGeometry.SetInt(Helpers.ShaderIDs.s_BlendSrcMode, (int)(_additive ? BlendMode.One : BlendMode.SrcAlpha));
+                _matGenerateWavesGeometry.SetInt(Helpers.ShaderIDs.s_BlendDstMode, (int)(_additive ? BlendMode.One : BlendMode.OneMinusSrcAlpha));
             }
 
             // Submit draws to create the Gerstner waves
@@ -664,7 +687,8 @@ namespace Crest
             {
                 if (i == -1) break;
                 _batches[i] = new GerstnerBatch(this, MinWavelength(i), i, _matGenerateWaves, _meshForDrawingWaves);
-                registered.Add(0, _batches[i]);
+                // Use the queue if local waves.
+                registered.Add(_meshForDrawingWaves != null ? _queue : int.MinValue, _batches[i]);
             }
         }